Move shortcutRepresentation to component
That way it can be used across the app more easily
